// src/components/Input/Input.d.ts
import * as React from 'react';
import { type VariantProps } from 'class-variance-authority';
declare const inputVariants: (props?: ({
    size?: "sm" | "md" | "lg" | null | undefined;
    tone?: "default" | "invalid" | "success" | null | undefined;
    hasLeft?: boolean | null | undefined;
    hasRight?: boolean | null | undefined;
} & import("class-variance-authority/types").ClassProp) | undefined) => string;
export interface InputProps extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 'size'>, VariantProps<typeof inputVariants> {
    leftIcon?: React.ReactNode;
    rightIcon?: React.ReactNode;
    hint?: string;
    errorText?: string;
    loading?: boolean;
    id?: string;
}
export declare const Input: React.ForwardRefExoticComponent<InputProps & React.RefAttributes<HTMLInputElement>>;
export {};